refactor: Move `set_query_timeout()`
- Change: `.utils.set_query_timeout()` -> `.set_query_timeout()`.
